Se separa la creación del fichero log con el envio en vivo de trazas directamente a Elasticsearch
